feat(section 9): completed section 09-go-type-system

+++ 09-go-type-system/exercises/02-the-type-problem/main.go
@@ -27,7 +27,7 @@ func main() {
 	// Change the following data types to the correct
 	// data types where appropriate.
 	var (
-		width  uint8
+		width  uint16
 		height uint16
 	)
